requirements.txt
pandas>=2.3.2
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.43
python-dotenv>=1.0.0